Fix ContinueDialog to work better after background thread callback to bot
Originally ContinueDialog was in the MainDialog and being called back after long-running tasks on the background thread like getting attractors, voids etc. But when I split the Scan actions into a sub-dialog, I was NPEs on clicking Scan xxx, value: storage can't be null simply just passing ConversationState over from VFatumbot to MainDialog. This approach of doing it in the Adapter seems to work better now.